refactoring vsComputer/page.tsx
